Move `domains/forwardings` to `domains/domain-forwardings`
